refactor: Add param count in use pagination hooks and styles adjustment in view product page, best sellers and new arrival component
