+++ client/src/pages/default/Home/components/NewArrivals/NewArrivals.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import { getProductsBySortAndOrder } from '../../../../library/services/product';
-import Card from '../../../../library/components/Card';
-import Pagination from '../../../../library/components/Pagination';
-import usePagination from '../../../../library/hooks/usePagination';
+import { getProductsBySortAndOrder } from '../../../../../library/services/product';
+import Card from '../../../../../library/components/Card';
+import Pagination from '../../../../../library/components/Pagination';
+import usePagination from '../../../../../library/hooks/usePagination';
 
 const NewArrivals = () => {
 	const [
@@ -20,7 +20,7 @@ const NewArrivals = () => {
 		previousPage,
 		nextPage,
 		dataResult,
-	} = usePagination(products);
+	} = usePagination(products, 4);
 
 	async function fetchProductsByCount(){
 		try {
@@ -44,15 +44,16 @@ const NewArrivals = () => {
 	}, []);
 
 	return (
-		<div className='my-5'>
+		<div className='my-10'>
 			<label className='text-2xl font-semibold'>New Arrivals</label>
 			<div className='flex'>
-				<div className='grid grid-cols-3 gap-4'>
+				<div className='grid grid-cols-4 gap-4'>
 					{
 						loading ? <h2>🌀 Loading....</h2> :
 						dataResult &&
 						dataResult.map((item) => (
 							<Card
+								imgContainerClass='relative flex h-60 overflow-hidden rounded-xl'
 								imgSrc={item.images}
 								title={item.title}
 								description={item.description}
