+++ client/src/pages/default/Home/components/NewArrivals/index.js
