+++ client/src/pages/default/ViewProduct/ViewProduct.jsx
@@ -30,13 +30,9 @@ const ViewProduct = ({ match }) => {
 		// eslint-disable-next-line
 	}, []);
 
-	const { title, description } = product || {};
-	const images = [
-		'https://media.istockphoto.com/id/1434947710/photo/businessman-headphones-and-laptop-webinar-in-office-with-coffee-on-table-video-call-or.jpg?s=1024x1024&w=is&k=20&c=NvC5p29pg1jBXw-IEzCTYg3Mv1A11k8BGVFqRw-DCDk=',
-		'https://media.istockphoto.com/id/1500285927/photo/young-woman-a-university-student-studying-online.jpg?s=1024x1024&w=is&k=20&c=CVhpekieDK_UB8vtEDw-dKKGWzDpsxcQt-XEQIkgm3Y=',
-		'https://via.placeholder.com/600x400',
-	];
+	const { title, description, brand, images } = product || {};
 
+	// console.log(product);
 	return (
 		<div className='w-full max-w-screen-xl mx-auto whitespace-pre-wrap break-words'>
 			{
@@ -46,7 +42,8 @@ const ViewProduct = ({ match }) => {
 						<div className='col-span-1'>
 							<Carousel images={images} />
 						</div>
-						<div className='col-span-1 mt-5'>
+						<div className='col-span-1 mx-5'>
+							<div className='text-gray-400 text-xs'>{brand}</div>
 							<label className='text-2xl font-bold'>{title}</label>
 							<p className='my-5'>{description}</p>
 							<button
